Move the button to a dedicated file for now.
