+++ packages/commonwealth/client/scripts/views/pages/discussions/ThreadCard/ThreadOptions/ThreadOptions.tsx
@@ -1,27 +1,14 @@
 import { pluralize } from 'helpers';
 import { GetThreadActionTooltipTextResponse } from 'helpers/threads';
-import { useFlag } from 'hooks/useFlag';
 import Thread from 'models/Thread';
-import React, {
-  Dispatch,
-  SetStateAction,
-  useCallback,
-  useMemo,
-  useState,
-} from 'react';
-import { useCreateThreadSubscriptionMutation } from 'state/api/trpc/subscription/useCreateThreadSubscriptionMutation';
-import { useDeleteThreadSubscriptionMutation } from 'state/api/trpc/subscription/useDeleteThreadSubscriptionMutation';
+import React, { Dispatch, SetStateAction } from 'react';
+import useUserStore from 'state/ui/user';
 import Permissions from 'utils/Permissions';
 import { downloadDataAsFile } from 'utils/downloadDataAsFile';
 import { SharePopover } from 'views/components/SharePopover';
 import { ViewUpvotesDrawerTrigger } from 'views/components/UpvoteDrawer';
 import { CWThreadAction } from 'views/components/component_kit/new_designs/cw_thread_action';
-import { useThreadSubscriptions } from 'views/pages/NotificationSettings/useThreadSubscriptions';
-import {
-  getCommentSubscription,
-  getReactionSubscription,
-  handleToggleSubscription,
-} from '../../helpers';
+import { ToggleThreadSubscribe } from 'views/pages/discussions/ThreadCard/ThreadOptions/ToggleThreadSubscribe';
 import { AdminActions, AdminActionsProps } from './AdminActions';
 import { ReactionButton } from './ReactionButton';
 import './ThreadOptions.scss';
@@ -70,96 +57,13 @@ export const ThreadOptions = ({
   setIsUpvoteDrawerOpen,
   editingDisabled,
 }: OptionsProps) => {
-  const [isSubscribed, setIsSubscribed] = useState(
-    thread &&
-      getCommentSubscription(thread)?.isActive &&
-      getReactionSubscription(thread)?.isActive,
-  );
-
   const isCommunityMember = Permissions.isCommunityMember(thread.communityId);
-
-  const enableKnockInAppNotifications = useFlag('knockInAppNotifications');
-
-  const doToggleSubscribeOld = useCallback(async () => {
-    if (!thread) {
-      return;
-    }
-
-    await handleToggleSubscription(
-      thread,
-      getCommentSubscription(thread),
-      getReactionSubscription(thread),
-      isSubscribed,
-      setIsSubscribed,
-    );
-  }, [isSubscribed, thread]);
 
   const handleDownloadMarkdown = () => {
     downloadDataAsFile(thread.plaintext, 'text/markdown', thread.title + '.md');
   };
 
-  const createThreadSubscriptionMutation =
-    useCreateThreadSubscriptionMutation();
-  const deleteThreadSubscriptionMutation =
-    useDeleteThreadSubscriptionMutation();
-
-  const threadSubscriptions = useThreadSubscriptions();
-
-  const hasThreadSubscriptionDefault = useMemo(() => {
-    const matching = (threadSubscriptions.data || []).filter(
-      (current) => current.thread_id === thread.id,
-    );
-
-    return matching.length > 0;
-  }, [thread.id, threadSubscriptions.data]);
-
-  const [hasThreadSubscriptionState, setHasThreadSubscriptionState] = useState<
-    boolean | undefined
-  >(undefined);
-
-  const hasThreadSubscription =
-    hasThreadSubscriptionState !== undefined
-      ? hasThreadSubscriptionState
-      : hasThreadSubscriptionDefault;
-
-  const doToggleSubscribe = useCallback(async () => {
-    if (hasThreadSubscription) {
-      await deleteThreadSubscriptionMutation.mutateAsync({
-        id: thread.id,
-        thread_ids: [thread.id],
-      });
-    } else {
-      await createThreadSubscriptionMutation.mutateAsync({
-        id: thread.id,
-        thread_id: thread.id,
-      });
-    }
-    setHasThreadSubscriptionState(!hasThreadSubscription);
-  }, [
-    createThreadSubscriptionMutation,
-    deleteThreadSubscriptionMutation,
-    hasThreadSubscription,
-    thread.id,
-  ]);
-
-  const handleToggleSubscribe = useCallback(
-    (e: React.MouseEvent) => {
-      async function doAsync() {
-        if (enableKnockInAppNotifications) {
-          await doToggleSubscribe();
-        } else {
-          await doToggleSubscribeOld();
-        }
-      }
-
-      // prevent clicks from propagating to discussion row
-      e.preventDefault();
-      e.stopPropagation();
-
-      doAsync().catch(console.error);
-    },
-    [doToggleSubscribe, doToggleSubscribeOld, enableKnockInAppNotifications],
-  );
+  const userStore = useUserStore();
 
   return (
     <>
@@ -211,23 +115,10 @@ export const ThreadOptions = ({
           {/* @ts-expect-error StrictNullChecks*/}
           <SharePopover linkToShare={shareEndpoint} buttonLabel="Share" />
 
-          {!enableKnockInAppNotifications && (
-            <CWThreadAction
-              action="subscribe"
-              label="Subscribe"
-              onClick={handleToggleSubscribe}
-              selected={!isSubscribed}
-              disabled={!isCommunityMember}
-            />
-          )}
-
-          {enableKnockInAppNotifications && (
-            <CWThreadAction
-              action="subscribe"
-              label="Subscribe"
-              onClick={handleToggleSubscribe}
-              selected={!hasThreadSubscription}
-              disabled={!isCommunityMember}
+          {userStore.id > 0 && (
+            <ToggleThreadSubscribe
+              thread={thread}
+              isCommunityMember={isCommunityMember}
             />
           )}
